fix: resolve 16 critical audit findings for v0.69.1 release

+++ frp-core/src/protocol.rs
@@ -13,13 +13,13 @@ pub async fn write_v1_frame<W: AsyncWriteExt + Unpin>(
     let payload = serde_json::to_vec(msg)
         .map_err(|e| crate::Error::Protocol(format!("serialize V1 msg: {e}")))?;
 
-    if payload.len() as i64 > V1_MAX_MSG_LENGTH {
+    if payload.len() as u64 > V1_MAX_MSG_LENGTH as u64 {
         return Err(crate::Error::Protocol("V1 message too large".into()));
     }
 
     let mut buf = Vec::with_capacity(V1_HEADER_LEN + payload.len());
     buf.push(type_byte);
-   buf.extend_from_slice(&(payload.len() as i64).to_be_bytes());
+   buf.extend_from_slice(&(payload.len() as u64).to_be_bytes());
    buf.extend_from_slice(&payload);
 
     tracing::trace!(
@@ -794,13 +794,9 @@ mod tests {
                 allow_users: None, bandwidth_limit: None, bandwidth_limit_mode: None,
                 annotations: None, metas: None, multiplexer: None, virtual_net: None,
                 proxy_protocol_version: None,
-                #[cfg(feature = "vnet")]
                 advertise_subnet: None,
-                #[cfg(feature = "vnet")]
                 vnet_ip: None,
-                #[cfg(feature = "vnet")]
                 vnet_netmask: None,
-                #[cfg(feature = "vnet")]
                 vnet_mtu: None,
             }),
             FrpMessage::NewProxyResp(msg::NewProxyResp {
